changed frametime type to int

+++ Super2DGameYo/code/Components/Animation.h
@@ -5,7 +5,7 @@ struct Animation
 {
 	Animation() = default;
 	float elapsed = 0.f;
-	float frameTime = 0.f;
+	int frameTime = 0;
 	std::vector<int> frames;
 	int currentFrame = 0;
 	entt::entity entity = entt::null;
